removed duplicate is_dicom_file
hello,

I
- removed the is_dicom_file function from hazenlib.init.
- moved the test_is_dicom_file to test_tools
- used os.join method

Let me know if this works.
