+++ tests/test_tools.py
@@ -1,4 +1,5 @@
 import unittest
+import os
 
 import pydicom
 
@@ -75,4 +76,36 @@ class TestShapeDetector(TestTools):
         shape_detector = hazen_tools.ShapeDetector(arr=arr)
         centre, size, angle = shape_detector.get_shape('rectangle')
         assert (centre, size, angle) == (self.cor2_rectangle_centre, self.cor2_rectangle_size, self.cor2_rectangle_angle)
+
+class Test_is_Dicom_file(unittest.TestCase):
+
+
+    def test_is_dicom(self):
+        folder = r".\tests\data\tools"
+        files = [os.path.join(folder, x) for x in os.listdir(folder)]
+
+
+        result = hazen_tools.is_dicom_file(files[1])
+        self.assertTrue(result)
+
+
+        result = hazen_tools.is_dicom_file(files[0])
+        self.assertFalse(result)
+
+    def test_is_dicom_yes(self):
+        folder = r".\tests\data\tools"
+        files = [os.path.join(folder, x) for x in os.listdir(folder)]
+        result = hazen_tools.is_dicom_file(files[1])
+        self.assertTrue(result)
+
+    def test_is_dicom_no(self):
+        folder = r".\tests\data\tools"
+        files = [os.path.join(folder, x) for x in os.listdir(folder)]
+        result = hazen_tools.is_dicom_file(files[0])
+        self.assertFalse(result)
+
+
+if __name__ == "__main__":
+    unittest.main()
+
